+++ MANIFEST.in
@@ -1,4 +1,5 @@
 include README.md
 include markdown_checker/inputs/*
 include markdown_checker/paths/*
+include markdown_checker/links/*
 include setup.cfg
